requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.2
discord.py==1.7.1
easydict==1.9
EasyProcess==0.3
entrypoint2==0.2.4
idna==2.10
keyboard==0.13.5
mss==6.1.0
multidict==5.1.0
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pypiwin32==223
pyscreenshot==3.0
pywin32==300
typing-extensions==3.7.4.3
urllib3==1.26.4
win10toast==0.9
yarl==1.6.3
